+++ blocks/news/news.js
@@ -49,9 +49,21 @@ export async function fetchData(type) {
   return data;
 }
 
+export async function getNewsData(newsCount) {
+  const isCuntryCodeZH = document.querySelector('html').getAttribute('lang') === 'en';
+  let data = [];
+  if (isCuntryCodeZH) {
+    data = await fetchData('china-news');
+  } else {
+    data = await fetchData('news');
+  }
+  if (newsCount) {
+    data = data.slice(0, newsCount);
+  }
+  return data;
+}
+
 export default async function decorate(block) {
-  //const isCuntryCodeZH = document.querySelector('html').getAttribute('lang') === 'zh';
-  const isCuntryCodeZH = 'zh';
   const config = readBlockConfig(block);
   placeholders = await fetchPlaceholders();
   const options = {
@@ -65,11 +77,7 @@ export default async function decorate(block) {
   options.activeFilters.set('year', '');
   options.activeFilters.set('page', 1);
 
-  if (isCuntryCodeZH === 'zh') {
-    options.data = await fetchData('china-news');
-  } else {
-    options.data = await fetchData('news');
-  }
+  options.data = await getNewsData();
 
   await createOverview(
     block,
